+++ packages/rust.sh
@@ -8,6 +8,11 @@ if (( $# > 0 )) && [ "$1" = 'description' ];then
     exit 0
 fi
 
+if [ -f "$(which cargo)" ];then
+    echo 'Already Installed'
+    exit 0
+fi
+
 sudo apt-get install -y curl
 curl https://sh.rustup.rs -sSf|sh -s -- -y
 source $HOME/.cargo/env
