+++ packages/thunderbird.sh
@@ -8,5 +8,10 @@ if (( $# > 0 )) && [ "$1" = 'description' ];then
     exit 0
 fi
 
+if [ -f "$(which thunderbird)" ];then
+    echo 'Already Installed'
+    exit 0
+fi
+
 sudo apt-get install -y thunderbird
